chore: moving current cmd package to cmd/v1
