--- cmd/all/all.go
@@ -1,6 +1,0 @@
-package all
-
-import (
-	_ "github.com/internetarchive/Zeno/cmd/get"
-	_ "github.com/internetarchive/Zeno/cmd/version"
-)
