+++ cmd/v1/all/all.go
@@ -1,0 +1,6 @@
+package all
+
+import (
+	_ "github.com/internetarchive/Zeno/cmd/v1/get"
+	_ "github.com/internetarchive/Zeno/cmd/v1/version"
+)
