+++ cmd/v1/get/url.go
@@ -3,7 +3,7 @@ package get
 import (
 	"net/url"
 
-	"github.com/internetarchive/Zeno/cmd"
+	"github.com/internetarchive/Zeno/cmd/v1"
 	"github.com/internetarchive/Zeno/config"
 	"github.com/internetarchive/Zeno/internal/pkg/frontier"
 	"github.com/sirupsen/logrus"
