+++ cmd/v1/version/deps.go
