+++ cmd/v1/version/version.go
@@ -1,7 +1,7 @@
 package version
 
 import (
-	"github.com/internetarchive/Zeno/cmd"
+	"github.com/internetarchive/Zeno/cmd/v1"
 	"github.com/internetarchive/Zeno/internal/pkg/utils"
 	"github.com/urfave/cli/v2"
 )
